% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bwt}
\alias{bwt}
\title{Burrows–Wheeler transform}
\usage{
bwt(string)
}
\arguments{
\item{string}{String that the transformation will be applied to.}
}
\description{
This function performs a Burrows–Wheeler transform of a sting input.
}
