% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{invbwt}
\alias{invbwt}
\title{Inverse Burrows–Wheeler transform}
\usage{
invbwt(string)
}
\arguments{
\item{string}{String that the inverse transformation will be applied to.}
}
\description{
This function performs an inverse of Burrows–Wheeler transform of a sting input.
}
